[incompat] update OpenSpliceTransport
